#1307 Refactoring started: changed config, separated logic from the jobs and extracted commonalities
