+++ spark-jobs/src/main/scala/za/co/absa/enceladus/conformance/interpreter/DynamicInterpreter.scala
@@ -104,7 +104,7 @@ object DynamicInterpreter {
         }
         rulesApplied += 1
         interpreter.conformanceRule match {
-          case Some(rule) => applyRuleCheckpoint(rule, conformedDf, progArgs.persistStorageLevel, explosionState.explodeContext)
+          case Some(rule) => applyRuleCheckpoint(rule, conformedDf, progArgs.jobConfig.persistStorageLevel, explosionState.explodeContext)
           case None       => conformedDf
         }
     })
@@ -266,7 +266,8 @@ object DynamicInterpreter {
     val fsUtils = new FileSystemVersionUtils(ictx.spark.sparkContext.hadoopConfiguration)
 
     val mappingTableDef = ictx.dao.getMappingTable(rule.mappingTable, rule.mappingTableVersion)
-    val mappingTablePath = PartitioningUtils.getPartitionedPathName(mappingTableDef.hdfsPath, ictx.progArgs.reportDate)
+    val mappingTablePath = PartitioningUtils.getPartitionedPathName(mappingTableDef.hdfsPath,
+      ictx.progArgs.jobConfig.reportDate)
     val mappingTableSize = fsUtils.getDirectorySizeNoHidden(mappingTablePath)
     (mappingTableSize / (1024 * 1024)).toInt
   }
