+++ spark-jobs/src/main/scala/za/co/absa/enceladus/conformance/interpreter/rules/CastingRuleInterpreter.scala
@@ -37,9 +37,9 @@ case class CastingRuleInterpreter(rule: CastingConformanceRule) extends RuleInte
   def conform(df: Dataset[Row])
              (implicit spark: SparkSession, explosionState: ExplosionState, dao: MenasDAO, progArgs: ConfCmdConfig): Dataset[Row] = {
     // Validate the rule parameters
-    RuleValidators.validateInputField(progArgs.datasetName, ruleName, df.schema, rule.inputColumn)
-    RuleValidators.validateOutputField(progArgs.datasetName, ruleName, df.schema, rule.outputColumn)
-    RuleValidators.validateSameParent(progArgs.datasetName, ruleName, rule.inputColumn, rule.outputColumn)
+    RuleValidators.validateInputField(progArgs.jobConfig.datasetName, ruleName, df.schema, rule.inputColumn)
+    RuleValidators.validateOutputField(progArgs.jobConfig.datasetName, ruleName, df.schema, rule.outputColumn)
+    RuleValidators.validateSameParent(progArgs.jobConfig.datasetName, ruleName, rule.inputColumn, rule.outputColumn)
 
     SchemaUtils.getFieldType(rule.inputColumn, df.schema)
       .foreach(dt => RuleValidators.validateTypeCompatibility(ruleName, rule.inputColumn, dt, rule.outputDataType))
