+++ spark-jobs/src/test/scala/za/co/absa/enceladus/conformance/interpreter/InterpreterSuite.scala
@@ -20,6 +20,7 @@ import org.json4s.native.JsonParser._
 import org.mockito.Mockito.{mock, when => mockWhen}
 import org.scalatest.{BeforeAndAfterAll, FunSuite}
 import za.co.absa.atum.model.ControlMeasure
+import za.co.absa.enceladus.common.JobCmdConfig
 import za.co.absa.enceladus.conformance.ConfCmdConfig
 import za.co.absa.enceladus.conformance.datasource.DataSource
 import za.co.absa.enceladus.dao.MenasDAO
@@ -49,7 +50,8 @@ class InterpreterSuite extends FunSuite with SparkTestBase with BeforeAndAfterAl
     spark.sessionState.conf.setConfString("co.za.absa.enceladus.confTest", "hello :)")
 
     implicit val dao: MenasDAO = mock(classOf[MenasDAO])
-    implicit val progArgs: ConfCmdConfig = ConfCmdConfig(reportDate = "2017-11-01", experimentalMappingRule = Option(useExperimentalMappingRule))
+    implicit val progArgs: ConfCmdConfig = ConfCmdConfig(experimentalMappingRule = Option(useExperimentalMappingRule),
+      jobConfig = JobCmdConfig(reportDate = "2017-11-01"))
     val enableCF = true
     val isCatalystWorkaroundEnabled = true
 
@@ -104,7 +106,8 @@ class InterpreterSuite extends FunSuite with SparkTestBase with BeforeAndAfterAl
     spark.enableControlMeasuresTracking("src/test/testData/_tradeData/2017/11/01/_INFO", "src/test/testData/_tradeOutput/_INFO")
 
     implicit val dao: MenasDAO = mock(classOf[MenasDAO])
-    implicit val progArgs: ConfCmdConfig = ConfCmdConfig(reportDate = "2017-11-01", experimentalMappingRule = Option(useExperimentalMappingRule))
+    implicit val progArgs: ConfCmdConfig = ConfCmdConfig(experimentalMappingRule = Option(useExperimentalMappingRule),
+      jobConfig = JobCmdConfig(reportDate = "2017-11-01"))
     val enableCF = true
     val isCatalystWorkaroundEnabled = true
