+++ spark-jobs/src/test/scala/za/co/absa/enceladus/conformance/interpreter/LiteralJoinMappingRuleTest.scala
@@ -17,6 +17,7 @@ package za.co.absa.enceladus.conformance.interpreter
 
 import org.mockito.Mockito.{mock, when => mockWhen}
 import org.scalatest.FunSuite
+import za.co.absa.enceladus.common.JobCmdConfig
 import za.co.absa.enceladus.conformance.ConfCmdConfig
 import za.co.absa.enceladus.conformance.datasource.DataSource
 import za.co.absa.enceladus.dao.MenasDAO
@@ -33,7 +34,7 @@ class LiteralJoinMappingRuleTest extends FunSuite with SparkTestBase with Logger
     val inputDf = spark.read.option("header", "true").csv("src/test/resources/interpreter/literalJoin/data")
     val mappingDf = spark.read.option("header", "true").csv("src/test/resources/interpreter/literalJoin/mapping")
 
-    implicit val progArgs: ConfCmdConfig = ConfCmdConfig(reportDate = "2018-03-23")
+    implicit val progArgs: ConfCmdConfig = ConfCmdConfig(jobConfig = JobCmdConfig(reportDate = "2018-03-23"))
     implicit val dao: MenasDAO = mock(classOf[MenasDAO])
     val enableCF = false
     val isCatalystWorkaroundEnabled = true
