+++ spark-jobs/src/test/scala/za/co/absa/enceladus/conformance/interpreter/fixtures/NestedStructsFixture.scala
@@ -22,6 +22,7 @@ import org.apache.spark.sql.{DataFrame, SaveMode}
 import org.mockito.Mockito.{mock, when => mockWhen}
 import org.scalatest.{BeforeAndAfterAll, Suite}
 import org.slf4j.{Logger, LoggerFactory}
+import za.co.absa.enceladus.common.JobCmdConfig
 import za.co.absa.enceladus.conformance.ConfCmdConfig
 import za.co.absa.enceladus.conformance.datasource.DataSource
 import za.co.absa.enceladus.dao.MenasDAO
@@ -38,7 +39,7 @@ trait NestedStructsFixture extends BeforeAndAfterAll with SparkTestBase {
   protected var standardizedDf: DataFrame = _
 
   implicit protected val dao: MenasDAO = mock(classOf[MenasDAO])
-  implicit protected val progArgs: ConfCmdConfig = ConfCmdConfig(reportDate = "2017-11-01")
+  implicit protected val progArgs: ConfCmdConfig = ConfCmdConfig(jobConfig = JobCmdConfig(reportDate = "2017-11-01"))
 
   protected val upperRule1 = UppercaseConformanceRule(order = 1, inputColumn = "strings.with_new_lines",
     controlCheckpoint = false, outputColumn = "strings.with_new_lines_upper")
