+++ spark-jobs/src/test/scala/za/co/absa/enceladus/standardization/StandardizationCobolEbcdicSuite.scala
@@ -19,6 +19,7 @@ import org.apache.spark.sql.DataFrame
 import org.apache.spark.sql.types.{StringType, StructField, StructType}
 import org.scalatest.mockito.MockitoSugar
 import org.scalatest.{Outcome, fixture}
+import org.slf4j.Logger
 import za.co.absa.enceladus.dao.MenasDAO
 import za.co.absa.enceladus.model.Dataset
 import za.co.absa.enceladus.standardization.fixtures.TempFileFixture
@@ -29,6 +30,9 @@ class StandardizationCobolEbcdicSuite extends fixture.FunSuite with SparkTestBas
   type FixtureParam = String
 
   private implicit val dao: MenasDAO = mock[MenasDAO]
+
+  private implicit val log: Logger = mock[Logger]
+  private val standardizationReader = new StandardizationReader(log)
 
   private val tmpFilePrefix = "cobol-fix-ebcdic-"
   private val tmpFileSuffix = ".dat"
@@ -71,7 +75,7 @@ class StandardizationCobolEbcdicSuite extends fixture.FunSuite with SparkTestBas
                                args: Array[String]
                               ): DataFrame = {
     val cmd: StdCmdConfig = StdCmdConfig.getCmdLineArguments(argumentsBase ++ args)
-    val cobolReader = StandardizationJob.getFormatSpecificReader(cmd, dataSet, schema.fields.length)
+    val cobolReader = standardizationReader.getFormatSpecificReader(cmd, dataSet, schema.fields.length)
     cobolReader
       .option("copybook_contents", copybook)
       .load(tmpFileName)
