+++ spark-jobs/src/test/scala/za/co/absa/enceladus/standardization/StandardizationJsonSuite.scala
@@ -18,6 +18,7 @@ package za.co.absa.enceladus.standardization
 import org.apache.spark.sql.types.{DataType, StructType}
 import org.scalatest.FunSuite
 import org.scalatest.mockito.MockitoSugar
+import org.slf4j.Logger
 import za.co.absa.enceladus.dao.MenasDAO
 import za.co.absa.enceladus.model.Dataset
 import za.co.absa.enceladus.standardization.interpreter.StandardizationInterpreter
@@ -30,6 +31,9 @@ import za.co.absa.enceladus.utils.udf.UDFLibrary
 class StandardizationJsonSuite extends FunSuite with SparkTestBase with MockitoSugar{
   private implicit val udfLibrary:UDFLibrary = new UDFLibrary()
 
+  private implicit val log: Logger = mock[Logger]
+  private val standardizationReader = new StandardizationReader(log)
+
   test("Reading data from JSON input, also such that don't adhere to desired schema") {
 
     implicit val dao: MenasDAO = mock[MenasDAO]
@@ -41,7 +45,7 @@ class StandardizationJsonSuite extends FunSuite with SparkTestBase with MockitoS
     val dataSet = Dataset("SpecialChars", 1, None, "", "", "SpecialChars", 1, conformance = Nil)
     val cmd = StdCmdConfig.getCmdLineArguments(args)
 
-    val csvReader = StandardizationJob.getFormatSpecificReader(cmd, dataSet)
+    val csvReader = standardizationReader.getFormatSpecificReader(cmd, dataSet)
 
     val baseSchema: StructType = DataType.fromJson(
       FileReader.readFileAsString("src/test/resources/data/standardization_json_suite_schema.json")
